+++ src/logging/log_macros.hpp
@@ -1,8 +1,8 @@
 #ifndef __LOGGING__LOG_MACROS_HPP__
 #define __LOGGING__LOG_MACROS_HPP__
 
-#include "log_manager.hpp"
 #include "log_entry_scope.hpp"
+#include "log_manager.hpp"
 #include "mstd/error.hpp"
 
 // clang-format off
@@ -21,12 +21,18 @@
 #ifndef __LOG_CATEGORY__
     #define LOG(level, message)                                                 \
         MSTD_COMPILE_FAIL(                                                      \
-            "LOG_CATEGORY must be defined before including log_manager.hpp"     \
+            "LOG_CATEGORY must be defined before including log_macros.hpp"      \
         );
 #else
     #define LOG(level, message) LOG_INTERNAL(level, __LOG_CATEGORY__, message)
 #endif
 // clang-format on
+
+#define LOG_TRACE(message)   LOG(LogLevel::Trace, message)
+#define LOG_DEBUG(message)   LOG(LogLevel::Debug, message)
+#define LOG_INFO(message)    LOG(LogLevel::Info, message)
+#define LOG_WARNING(message) LOG(LogLevel::Warning, message)
+#define LOG_ERROR(message)   LOG(LogLevel::Error, message)
 
 // clang-format off
 #define LOG_ENTRY                        \