+++ src/ui/menu_bar/debug_menu.cpp
@@ -1,0 +1,25 @@
+#include "debug_menu.hpp"
+
+#include <QMainWindow>
+#include <QMenu>
+
+namespace ui
+{
+
+    DebugMenu::DebugMenu(QMainWindow& mainWindow)
+        : QObject{&mainWindow}, _mainWindow{mainWindow}
+    {
+    }
+
+    void DebugMenu::build(QMenu* parentMenu)
+    {
+        _debugSlotsAction = parentMenu->addAction("Debug Slots");
+        connect(
+            _debugSlotsAction,
+            &QAction::triggered,
+            this,
+            &DebugMenu::requestDebugSlots
+        );
+    }
+
+}   // namespace ui