+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All changes and updates, that are relevant for a user will be documented here
 ### UI
 
 - Main window always opens in full screen mode
+- Add Dialog to set Debug flags for logging
 
 ### Error Handling